fixed routing, also for search/result
search is moved to own place, the search route changes based on if you have search or not, if you press the search button and are already in /result it will redirect to search.
home and settings added, only placeholder for now.
removed link to "home" on title / top of app